AP is a test of its own; ESP-NOW runs station-only; FTM signed, 11b kept
The "ftm resp" switch becomes "ap": the SoftAP as a comms test in its own
right, and the FTM responder. Without it the driver runs STA only, so an
espnow-only test exists again - the earlier "STA power-saves" reason for
always-APSTA was wrong (esp_now's wake window defaults to always awake).
Protocols are back to 11b (+LR); the 11g/n added for FTM is gone.

FTM: distances are signed (dist_est goes below zero uncalibrated and was
wrapping to kilometres), the name is the only filter (the responder bit is
counted, not required), and the 0x04 notification now always goes while
FTM is on, opening with what the last scan found. Report version 9.

+++ main/rt_log.c
@@ -320,7 +320,7 @@ void rt_log_rx(uint32_t node, int chan, int8_t ptx, uint32_t seq, uint32_t gap, 
     portEXIT_CRITICAL_SAFE(&s_mux);
 }
 
-void rt_log_ftm(uint32_t node, uint8_t status, uint16_t dist_dm, int8_t rssi)
+void rt_log_ftm(uint32_t node, uint8_t status, int16_t dist_dm, int8_t rssi)
 {
     if (s_ring == NULL) {
         return;
@@ -333,7 +333,7 @@ void rt_log_ftm(uint32_t node, uint8_t status, uint16_t dist_dm, int8_t rssi)
     put8(dt);
     put24(node);
     put8(status);
-    put16(dist_dm);
+    put16((uint16_t)dist_dm);
     put8((uint8_t)rssi);
     portEXIT_CRITICAL_SAFE(&s_mux);
 }
